+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format of this changelog is based on
 ### Fixed
 
   - `launch!` without rounding now has the correct gap behind the pad
+  - `terminate!` with `initial=true` appends the termination before the `Path` start as documented (previously incorrectly kept `p0(path)` constant, shifting the rest of the `Path` forward)
+  - `terminate!` with rounding on a curve is still drawn as straight but keeps the full underlying segment (previously consumed some turn angle to replace with straight segment including rounding length)
 
 ## 1.0.0 (2025-02-27)